Refactor home page product discovery layout
